+++ sprockets/mixins/mediatype/type_info.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import decimal
+import ipaddress
 import typing
 import uuid
 
@@ -40,7 +41,8 @@ class SupportsSettings(Protocol):
 Serializable = typing.Union[SupportsIsoFormat, None, bool, bytearray, bytes,
                             float, int, memoryview, str, typing.Mapping,
                             typing.Sequence, typing.Set, uuid.UUID,
-                            decimal.Decimal, SupportsDataclassFields]
+                            decimal.Decimal, SupportsDataclassFields,
+                            ipaddress.IPv4Address, ipaddress.IPv6Address]
 """Types that can be serialized by this library.
 
 This is the set of types that